@@ -1,6 +1,8 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
-import "../src/index.css";
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import App from './App';
+import '../src/index.css';
 
-ReactDOM.render(<App />, document.getElementById("root"));
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(<App />);
